Added input manager for the keyboard and mouse
Also added a callback object to store pointers to useful things used in callback functions.

Messing around with shininess settings, still not set on them.

Also added a variable number of point sources in the fragment shader so I don't need to keep changing that as I mess with the number of point sources.